+++ stellar_sdk/helpers.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import Union
 
 from .fee_bump_transaction_envelope import FeeBumpTransactionEnvelope
@@ -29,6 +30,10 @@ def parse_transaction_envelope_from_xdr(
     :raises: :exc:`ValueError <stellar_sdk.exceptions.ValueError>` - XDR is neither :py:class:`TransactionEnvelope <stellar_sdk.transaction_envelope.TransactionEnvelope>`
         nor :py:class:`FeeBumpTransactionEnvelope <stellar_sdk.fee_bump_transaction_envelope.FeeBumpTransactionEnvelope>`
     """
+    warnings.warn(
+        "This function is deprecated. Use `TransactionBuilder.from_xdr` instead.",
+        DeprecationWarning,
+    )
     if FeeBumpTransactionEnvelope.is_fee_bump_transaction_envelope(xdr):
         return FeeBumpTransactionEnvelope.from_xdr(xdr, network_passphrase)
     return TransactionEnvelope.from_xdr(xdr, network_passphrase)